pyproject.toml
[tool.poetry]
package-mode = false
authors      = ["dtx1007 <148854108+dtx1007@users.noreply.github.com>"]

[tool.poetry.dependencies]
python    = "3.11.13"
torch     = [
    {version = "^2.7.1", markers = "sys_platform == 'darwin' and extra == 'gpu' and extra != 'cpu'"},
    {version = "^2.7.1", source="torch_cuda121", markers = "sys_platform == 'linux' and extra == 'gpu' and extra != 'cpu'"},
    {version = "^2.7.1", source="torch_cuda121", markers = "sys_platform == 'win32' and extra == 'gpu' and extra != 'cpu'"},
    {version = "^2.7.1", markers = "sys_platform == 'darwin' and extra == 'cpu' and extra != 'gpu'"},
    {version = "^2.7.1", source="torchcpu", markers = "sys_platform == 'linux' and extra == 'cpu' and extra != 'gpu'"},
    {version = "^2.7.1", source="torchcpu", markers = "sys_platform == 'win32' and extra == 'cpu' and extra != 'gpu'"},
]
numpy         = "2.2.6"
pandas        = "^2.3.0"
scikit-learn  = "^1.7.0"
matplotlib    = "^3.10.3"
seaborn       = "^0.13.2"
xgboost       = "^3.0.2"
androguard    = "^4.1.3"
ppdeep        = "^20200505"
python-dotenv = "^1.1.0"
optuna = "^4.3.0"
streamlit = "^1.46.0"
numba = "^0.61.2"
shap = "^0.48.0"
umap-learn = "^0.5.7"
charset-normalizer = "^3.4.2"
jupyter = "^1.1.1"

[[tool.poetry.source]]
name     = "torchcpu"
url      = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[[tool.poetry.source]]
name     = "torch_cuda121"
url      = "https://download.pytorch.org/whl/cu126"
priority = "explicit"

[tool.poetry.extras]
cpu = ["torch"]
gpu = ["torch"]

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
